parser exchange progress in floats

+++ CPP09/ex00/BitcoinExchange/BitcoinExchange.hpp
@@ -35,7 +35,7 @@ using std::cerr;
 class BitcoinExchange
 {
 	public:
-		std::map<Date, float>	dicc;
+		std::map<Date, double>	dicc;
 
 		//CONSTRUCTORS
 		//-------------------------------------------------------
